fix: corner cases when send called while closed
Fixes problems caused by the app using Strophe trying to send
while the WebSocket is closed. That would put packets in the queue
and send them while doing the resume operation and breaking
the XMPP stream (server replies with unsupported stanza for example).
This was fixed by clearing the _data buffer on the Strophe.Connection.

Fixes a case where the stream management plugin would request acknowledgment
after the WebSocket died.

Added safety checks for the connection status here and there.